Server2/integrated-oreo-project/requirements.txt
Django>=3.2.0
gunicorn>=20.1.0
psycopg2-binary>=2.9.1
python-dotenv>=0.19.0
django-extensions>=3.1.0
django-allauth>=0.54.0
Pillow>=9.0.0
requests>=2.31.0
djangorestframework>=3.14.0
django-imagekit>=4.1.0
pilkit>=2.0
